+++ bin/jmespath.rb
@@ -6,6 +6,6 @@ require 'jmespath'
 require 'json'
 
 expression = ARGV[0]
-json = JSON.load(STDIN.read)
+json = JSON.parse(STDIN.read)
 
 $stdout.puts(JSON.dump(JMESPath.search(expression, json)))